+++ Tests/SentryTests/Helper/SentryEnabledFeaturesBuilderTests.swift
@@ -80,6 +80,7 @@ final class SentryEnabledFeaturesBuilderTests: XCTestCase {
         XCTAssertEqual(result, [])
     }
 
+    @available(*, deprecated, message: "The test is marked as deprecated to silence the deprecation warning of the tested property.")
     func testEnableExperimentalViewRenderer_isEnabled_shouldAddFeature() throws {
 #if os(iOS)
         // -- Arrange --
@@ -92,6 +93,40 @@ final class SentryEnabledFeaturesBuilderTests: XCTestCase {
 
         // -- Assert --
         XCTAssert(features.contains("experimentalViewRenderer"))
+#else
+        throw XCTSkip("Test not supported on this platform")
+#endif
+    }
+
+    func testEnableViewRendererV2_isEnabled_shouldAddFeature() throws {
+#if os(iOS)
+        // -- Arrange --
+        let options = Options()
+
+        options.sessionReplay.enableViewRendererV2 = true
+
+        // -- Act --
+        let features = SentryEnabledFeaturesBuilder.getEnabledFeatures(options: options)
+
+        // -- Assert --
+        XCTAssertTrue(features.contains("experimentalViewRenderer"))
+#else
+        throw XCTSkip("Test not supported on this platform")
+#endif
+    }
+
+    func testEnableViewRendererV2_isNotEnabled_shouldAddFeature() throws {
+#if os(iOS)
+        // -- Arrange --
+        let options = Options()
+
+        options.sessionReplay.enableViewRendererV2 = false
+
+        // -- Act --
+        let features = SentryEnabledFeaturesBuilder.getEnabledFeatures(options: options)
+
+        // -- Assert --
+        XCTAssertFalse(features.contains("experimentalViewRenderer"))
 #else
         throw XCTSkip("Test not supported on this platform")
 #endif