+++ CHANGELOG.md
@@ -2,6 +2,16 @@
 
 ## Unreleased
 
+> ![Important]
+> This version makes the new view renderer used by Session Replay generally available.
+> You can disable it by setting the option `options.sessionReplay.enableViewRendererV2` to `false`.
+>
+> In case you are noticing issues with the new view renderer, please report them on [GitHub](https://github.com/getsentry/sentry-cocoa).
+
+## Improvements
+
+- GA of new session replay view renderer (#5054)
+
 ### Fixes
 
 - Crash in setMeasurement when name is nil (#5064)